inference-server/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
ultralytics>=8.3.0
opencv-python-headless>=4.10.0.84
pillow>=10.4.0
numpy>=1.26.0
torch>=2.5.0
torchvision>=0.20.0
